// src/app/shared/interfaces/surveyinfo.ts
export interface SurveyInfo {
    survey_info_id: number;
    run_name: string;
    run_number: number;
    survey_run_in: string;
    minimum_id: string;
    north_reference: string;
    survey_calculation_method: string;
    map_zone: string;
    geodetic_system: string;
    geodetic_datum: string;
    start_depth: number;
    tag_depth: number;
    proposal_direction: number;
    job_number: string;
    type_of_tool: number;
    survey_type: number;
    hole_section: number;
}